requirements.txt
feedparser==5.1.3
gitpython==0.3.5
pyyaml==3.11
python-slugify==0.1.0
